change singly-constrained function parameter (comments)
Since in most paper the attractiveness parameter is written as Wj
instead of Dj.

Nevertheless, check the references again in order to clarify the terms
and to concisely describe them in the documentation as well as in the
paper

starting point for this can be: Rihll, T.E., Wilson, A.G., 1988. Modelling Settlement Structures. Working Paper, University Leeds 516.